Move QueryConverter into persistence package.
