+++ src/logicblocks/event/persistence/postgres/converter/__init__.py
@@ -1,10 +1,9 @@
-from ..types import ClauseConverter as ClauseConverter
-from ..types import QueryConverter as QueryConverter
 from .clause import FilterClauseConverter as FilterClauseConverter
 from .clause import KeySetPagingClauseConverter as KeySetPagingClauseConverter
 from .clause import OffsetPagingClauseConverter as OffsetPagingClauseConverter
 from .clause import SortClauseConverter as SortClauseConverter
 from .clause import TypeRegistryClauseConverter as TypeRegistryClauseConverter
+from .delegate import DelegatingQueryConverter as DelegatingQueryConverter
 from .query import LookupQueryConverter as LookupQueryConverter
 from .query import SearchQueryConverter as SearchQueryConverter
 from .query import TypeRegistryQueryConverter as TypeRegistryQueryConverter
