+++ src/logicblocks/event/processing/broker/subscribers/stores/state/postgres.py
@@ -6,15 +6,7 @@ from psycopg.rows import dict_row
 from psycopg.types.json import Jsonb
 from psycopg_pool import AsyncConnectionPool
 
-from logicblocks.event.persistence.postgres import (
-    ConnectionSettings,
-    ConnectionSource,
-    ParameterisedQuery,
-    TableSettings,
-)
-from logicblocks.event.projection.store.adapters import (
-    PostgresQueryConverter,
-)
+import logicblocks.event.persistence.postgres as postgres
 from logicblocks.event.query import (
     FilterClause,
     Operator,
@@ -30,8 +22,8 @@ from .base import EventSubscriberState, EventSubscriberStateStore
 
 def insert_query(
     subscriber: EventSubscriberState,
-    table_settings: TableSettings,
-) -> ParameterisedQuery:
+    table_settings: postgres.TableSettings,
+) -> postgres.ParameterisedQuery:
     subscription_requests_jsonb = Jsonb(
         [
             subscription_request.serialise()
@@ -67,8 +59,8 @@ def insert_query(
 
 def delete_query(
     key: EventSubscriber,
-    table_settings: TableSettings,
-) -> ParameterisedQuery:
+    table_settings: postgres.TableSettings,
+) -> postgres.ParameterisedQuery:
     return (
         sql.SQL(
             """
@@ -85,8 +77,8 @@ def delete_query(
 
 def heartbeat_query(
     subscriber: EventSubscriberState,
-    table_settings: TableSettings,
-) -> ParameterisedQuery:
+    table_settings: postgres.TableSettings,
+) -> postgres.ParameterisedQuery:
     return (
         sql.SQL(
             """
@@ -107,8 +99,8 @@ def heartbeat_query(
 
 def purge_query(
     cutoff_time: datetime,
-    table_settings: TableSettings,
-) -> ParameterisedQuery:
+    table_settings: postgres.TableSettings,
+) -> postgres.ParameterisedQuery:
     return (
         sql.SQL(
             """
@@ -128,14 +120,14 @@ class PostgresEventSubscriberStateStore(EventSubscriberStateStore):
         self,
         *,
         node_id: str,
-        connection_source: ConnectionSource,
+        connection_source: postgres.ConnectionSource,
         clock: Clock = SystemClock(),
-        table_settings: TableSettings = TableSettings(
+        table_settings: postgres.TableSettings = postgres.TableSettings(
             table_name="subscribers"
         ),
-        query_converter: PostgresQueryConverter | None = None,
+        query_converter: postgres.QueryConverter | None = None,
     ):
-        if isinstance(connection_source, ConnectionSettings):
+        if isinstance(connection_source, postgres.ConnectionSettings):
             self._connection_pool_owner = True
             self.connection_pool = AsyncConnectionPool[AsyncConnection](
                 connection_source.to_connection_string(), open=False
@@ -152,7 +144,7 @@ class PostgresEventSubscriberStateStore(EventSubscriberStateStore):
             query_converter
             if query_converter is not None
             else (
-                PostgresQueryConverter(table_settings=table_settings)
+                postgres.QueryConverter(table_settings=table_settings)
                 .with_default_clause_converters()
                 .with_default_query_converters()
             )
