+++ src/logicblocks/event/projection/store/adapters/postgres/__init__.py
@@ -1,11 +1,3 @@
 from .adapter import (
     PostgresProjectionStorageAdapter as PostgresProjectionStorageAdapter,
 )
-from .converter import (
-    PostgresQueryConverter as PostgresQueryConverter,
-)
-
-__all__ = [
-    "PostgresProjectionStorageAdapter",
-    "PostgresQueryConverter",
-]
